feature: change to use Composable in Box
